LayoutTests/http/tests/resourceLoadStatistics/cookie-deletion-expected.txt

CONSOLE MESSAGE: line 117: Test is beginning. document.location.hash is empty.
CONSOLE MESSAGE: line 123: About to call runTest() for the first time.
CONSOLE MESSAGE: line 62: step1. About to set a cookie
CONSOLE MESSAGE: line 127: About to call runTest() after cookie was set. document hash is #step2
CONSOLE MESSAGE: line 67: step2. About to open an iFrame to test for third party cookie access (should be successful)
CONSOLE MESSAGE: line 73: step3. About to classify localhost as prevalent
CONSOLE MESSAGE: line 77: step3. In the callback for testRunner.setStatisticsPrevalentResource
CONSOLE MESSAGE: line 84: step4. About to open an iFrame to test for third party cookie access (should not be successful)
CONSOLE MESSAGE: line 90: step5. About to open an iFrame to try to set a cookie as a third party (should fail)
CONSOLE MESSAGE: line 96: step6. About to open an iFrame and fireDataModificationHandlerAndContinue
CONSOLE MESSAGE: line 40: In fireDataModificationHandlerAndContinue
CONSOLE MESSAGE: line 45: Calling statisticsProcessStatisticsAndDataRecords
CONSOLE MESSAGE: line 42: In callback function for installStatisticsDidScanDataRecordsCallback
CONSOLE MESSAGE: line 102: step7. About to open an iFrame and setAsNonPrevalentAndContinue
CONSOLE MESSAGE: line 50: In setAsNonPrevalentAndContinue
CONSOLE MESSAGE: line 52: In callback function for setStatisticsPrevalentResource
CONSOLE MESSAGE: line 108: step8. About to open an iFrame and confirm third party has no cookie access
CONSOLE MESSAGE: line 25: in finishTest
Test for partitioned and unpartitioned cookie deletion.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame2-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame3-->'
--------
Try to set third party cookie.


--------
Frame: '<!--frame4-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame5-->'
--------
After removal, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame6-->'
--------
After declassified, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:
